Collectd+InfluxDb-RMON Replacement(ALL METRICS) P1
This update introduces a non-reboot patch example script for a service
whose patch modified its service file thereby requiring a system
daemon-reload as a precursor to the service process restart.
This new RPM was introduced in support of no-reboot patching of the
'influxdb' and 'collectd' processes ; so those are the default processes
used in this script.
Note that there is nothing that prevents a designer from using this as a
more general example script for any service that requires the same
patching operations after replacing the default process names with their
own.

Change-Id: I9f261509f16490f986722b47b0f92db4ef735609
Signed-off-by: Jack Ding <jack.ding@windriver.com>