Patch Signature check failed after promotion.
Problem: modify_patch.sh was signature unaware.
When it was used to promote a patch to 'rel' status,
it would be re-assembled with a 'dev' signing key,
regardless of which key was used to sign the original
patch.

Solution: Make modify_patch.sh aware of signatures,
and force it to try and use the same signature as was
already in use.  This may require it to submit the patch
to the signing server.  If signing failes, the overall
patch modification will fail.

Change-Id: I5319062b910259f4eaff3df1504216f09536c84e
Signed-off-by: Scott Little <scott.little@windriver.com>