stx-update: API ref doc content added.
I added the api-ref-sysinv-v1-update file, which is content
for the API reference manual to the api-ref/source dir.
This represents the converted old-style files to the newer
OpenStack supported RST files.

I updated the index.rst file to include the new .rst file
so that the api-ref document can build.

Change-Id: I683e67e07c25e74f2cebc13517609c7876d9746e
Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>