Adding unit tests to tsconfig and clean it up a bit
Adding simple unit tests to tsconfig.
- test the behaviour when build.info is not present
- test the behaviour when platform.conf is not present
- test the behaviour when both files are found

This works in py27 and py35.
Coverage for tsconfig is 100%

Added requirements.txt and test-requirements.txt

Added a setup.cfg file so that the wheel can be
built in universal mode (python 2 and 3)

Properly set some of the setup.cfg variables.

These unit tests are not currently wired into zuul.
That will be done in a subsequent review.

Story: 2004515
Task: 29783
Change-Id: Ifb4f5403592ccc9223659bb188055ffeccc113b5
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>