Protect against generating empty patch
If a user attempts to generate a patch without specifying
any packages, raise an exception to avoid malformed patches.

Additionally, very small files (like a patch with no packages)
appear to be handled differently by the API handler, passing
a cStringIO object for small files and a file object for files
that are about 1K or larger. For robustness, the upload
handler is updated to do an isinstance(file) check in order
to protect against an unhandled exception for small files.

Change-Id: Ie1e730a183a27850ba8567f31aea603459b59d45
Closes-Bug: 1803022
Signed-off-by: Don Penney <don.penney@windriver.com>